+++ R/dynamite.R
@@ -536,38 +536,93 @@ sampling_info <- function(dformulas, verbose, debug, backend) {
 #' @noRd
 check_stan_args <- function(dots, verbose, backend) {
   dots_names <- names(dots)
+  common_args <- c(
+    "chains", "thin", "seed", "init", "show_messages", "refresh", "save_warmup"
+  )
+  rstan_args <- c(
+    "pars", "iter", "warmup", "check_data", "sample_file", "diagnostic_file",
+    "verbose", "algorithm", "control", "include", "open_progress", "cores",
+    "chain_id", "init_r", "test_grad", "append_samples", "enable_random_init"
+  )
+  cmdstanr_args <- c(
+    "save_latent_dynamics", "output_dir", "output_basename", "sig_figs",
+    "parallel_chains", "chain_ids", "threads_per_chain", "opencl_ids",
+    "iter_warmup", "iter_sampling", "max_treedepth", "adapt_engaged",
+    "adapt_delta", "step_size", "metric", "metric_file", "inv_metric",
+    "init_buffer", "term_buffer", "window", "fixed_param", "show_exceptions",
+    "diagnostics"
+  )
+  # cores not included here as it will be converted
+  cmdstanr_deprecated_args <- c(
+    "num_cores", "num_chains", "num_warmup", "num_samples",
+    "validate_csv", "save_extra_diagnostics", "max_depth", "stepsize"
+  )
   args <- ifelse_(
     identical(backend, "rstan"),
+    c(common_args, rstan_args),
+    c(common_args, cmdstanr_args)
+  )
+  cmdstanr_to_rstan <- c(
+    `parallel_chains` = "cores",
+    `iter_warmup` = "warmup",
+    `iter_sampling` = "iter"
+  )
+  rstan_to_cmdstanr <- c(
+    `cores` = "parallel_chains",
+    `warmup` = "iter_warmup",
+    `iter` = "iter_sampling"
+  )
+  from <- c(`cmdstanr` = "rstan", `rstan` = "cmdstanr")[backend]
+  to <- c(`cmdstanr` = "rstan", `rstan` = "cmdstanr")[from]
+  original_args <- ifelse_(
+    identical(backend, "rstan"),
+    which(dots_names %in% names(cmdstanr_to_rstan)),
+    which(dots_names %in% names(rstan_to_cmdstanr))
+  )
+  conversion <- ifelse_(
+    identical(backend, "rstan"),
+    cmdstanr_to_rstan,
+    rstan_to_cmdstanr
+  )
+  m <- match(dots_names, names(conversion), nomatch = 0L)
+  converted_args <- unname(conversion[m[m > 0L]])
+  duplicate_args <- converted_args[converted_args %in% dots_names]
+  stopifnot_(
+    length(duplicate_args) == 0L,
     c(
-      "pars", "chains", "iter", "warmup", "thin", "seed", "init", "check_data",
-      "sample_file", "diagnostic_file", "verbose", "algorithm", "control",
-      "include", "cores", "open_progress", "show_messages", "chain_id",
-      "init_r", "test_grad", "append_samples", "refresh", "save_warmup",
-      "enable_random_init"
-    ),
-    c(
-      "seed", "refresh", "init", "save_latent_dynamics", "output_dir",
-      "output_basename", "sig_figs", "chains", "parallel_chains", "chain_ids",
-      "threads_per_chain", "opencl_ids", "iter_warmup", "iter_sampling",
-      "save_warmup", "thin", "max_treedepth", "adapt_engaged", "adapt_delta",
-      "step_size", "metric", "metric_file", "inv_metric", "init_buffer",
-      "term_buffer", "window", "fixed_param", "show_messages",
-      "show_exceptions", "diagnostics", "cores", "num_cores", "num_chains",
-      "num_warmup", "num_samples", "validate_csv", "save_extra_diagnostics",
-      "max_depth", "stepsize"
+      "Conflict in argument syntax conversion
+       from {.pkg {from}} to {.pkg {to}}.",
+      `x` = "Argument{?s} {.arg {duplicate_args}} {?has/have} been
+             multiply specified."
     )
   )
+  dots_names[original_args] <- converted_args
+  names(dots)[original_args] <- converted_args
+  if (identical(backend, "cmdstanr") ) {
+    valid_args <- !dots_names %in% cmdstanr_deprecated_args
+    deprecated_args <- dots_names[!valid_args]
+    if (verbose && any(deprecated_args)) {
+      warning_(
+        "{cli::qty(deprecated_args)}
+        Argument{?s} {.arg {deprecated_args}} passed to {.pkg {backend}}
+        sampling function {cli::qty(deprecated_args)}{?is/are} deprecated and
+        will be ignored."
+      )
+    }
+    dots_names <- dots_names[valid_args]
+    dots <- dots[valid_args]
+  }
   valid_args <- dots_names %in% args
   invalid_args <- dots_names[!valid_args]
-  dots <- dots[valid_args]
   if (verbose && any(!valid_args)) {
     warning_(
       "{cli::qty(invalid_args)}
-       Argument{?s} {.arg {invalid_args}} passed to {backend} sampling function
-       {cli::qty(invalid_args)}{?is/are} not recognized and will be ignored."
+      Argument{?s} {.arg {invalid_args}} passed to {.pkg {backend}} sampling
+      function {cli::qty(invalid_args)}{?is/are} not recognized and will be
+      ignored."
     )
   }
-  dots
+  dots[valid_args]
 }
 
 #' Count The Number of Random Effects in a `dynamiteformula`
@@ -617,7 +672,6 @@ count_random_effects <- function(dformula, data) {
   }
   M
 }
-
 
 #' Remove Redundant Parameters When Using `rstan`
 #'